[MISC] move type_traits/function to utility
