+++ include/seqan3/alignment/pairwise/alignment_algorithm.hpp
@@ -32,11 +32,11 @@
 #include <seqan3/core/simd/simd.hpp>
 #include <seqan3/core/simd/simd_traits.hpp>
 #include <seqan3/core/simd/view_to_simd.hpp>
-#include <seqan3/core/type_traits/function.hpp>
 #include <seqan3/range/container/aligned_allocator.hpp>
 #include <seqan3/range/views/drop.hpp>
 #include <seqan3/range/views/get.hpp>
 #include <seqan3/range/views/take.hpp>
+#include <seqan3/utility/type_traits/function_traits.hpp>
 
 namespace seqan3::detail
 {
